tree: move tests that pull in server to a different package
This adds the tree/eval_test package, which has dependencies on server
and colexec. The tree_test package itself doesn't have to have these
heavy deps anymore, which will speed up debugging cycles.

Release note: None
